feat: consolidate terminal output behind a stateful Terminal facade
Moves all ANSI / tabled display code into `cli/src/terminal/`
(Rust-2024 layout, no mod.rs):

- `terminal.rs` is the module root with `Terminal`, `OutcomeDisplay`,
  and the ANSI colour constants.
- `terminal/banner.rs` and `terminal/table.rs` migrate from the cli
  crate root.
- `terminal/spinner.rs` is new: a tiny `Spinner` struct that owns the
  indeterminate progress bar + colour-aware no-op path, replacing the
  two loose `start_spinner` / `finish_spinner` helpers.

`Terminal` is stateful across one command invocation (owns
`suite_rows` + the active `Spinner`). Callers use a
start_run / stop_run / finish bracket:

    let mut terminal = Terminal::new(total);
    for (index, config) in configs.iter().enumerate() {
        terminal.start_run(index + 1, config);
        let result = … ;
        terminal.stop_run(&result);
    }
    terminal.finish();

`simulate.rs` no longer manages spinner lifecycle, builds `SuiteRow`s,
or decides whether to print the aggregate table. Outcome classification
is unchanged — commands read `result.outcome` for the exit-code path.

The temporary `From<&RunResult<C>>` constructor on `Terminal` is
dropped: per-result rendering is now a method on the stateful type,
not an on-the-fly wrapper.

`BannerPrinter` is re-exported from the `terminal` module; the thin
`print_banner` wrapper was dropped in favour of direct construction
(`terminal::BannerPrinter::new(…).print()`).

Refs #61, tracked by #66.
