+++ crates/cli/src/lib.rs
@@ -2,9 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod args;
-pub mod banner;
 pub mod commands;
 pub mod report;
-pub mod reporter;
-pub mod table;
+pub mod terminal;
 pub mod tracing;
