--- crates/cli/src/reporter.rs
@@ -1,236 +1,0 @@
-// Copyright (c) Mysten Labs, Inc.
-// SPDX-License-Identifier: Apache-2.0
-
-use std::{io::IsTerminal, time::Duration};
-
-use dag::metrics::{AggregateMetrics, Outcome, RunResult};
-use eyre::{Result, eyre};
-use indicatif::{ProgressBar, ProgressStyle};
-use simulator::{SimulationConfig, SimulationRunner};
-
-use crate::{
-    banner::BannerPrinter,
-    table::{self, ConfigRow, ReplicaRow, SuiteRow},
-};
-
-/// Display helpers for the plain [`Outcome`] enum. Kept as an extension trait in the
-/// renderer rather than inherent methods on the domain type; migrates into `terminal/`
-/// along with the rest of the renderer in #61.
-pub trait OutcomeDisplay {
-    /// Single-character glyph, no colour. Usable standalone (e.g. inside a table cell).
-    fn glyph(&self) -> &'static str;
-    /// Fully-formatted badge line. When `color=true`, wraps the glyph + prose in ANSI
-    /// colour escapes; when `false`, returns `"LABEL: prose"`.
-    fn badge(&self, color: bool) -> String;
-}
-
-impl OutcomeDisplay for Outcome {
-    fn glyph(&self) -> &'static str {
-        match self {
-            Outcome::Pass => "✓",
-            Outcome::NoProgress => "⚠",
-            Outcome::Diverged => "✗",
-        }
-    }
-
-    fn badge(&self, color: bool) -> String {
-        let prose = match self {
-            Outcome::Pass => "Commits consistent across all replicas",
-            Outcome::NoProgress => "Safe but no leader was committed",
-            Outcome::Diverged => "Commits DIVERGED across replicas",
-        };
-        if color {
-            let color_code = match self {
-                Outcome::Pass => GREEN,
-                Outcome::NoProgress => YELLOW,
-                Outcome::Diverged => RED,
-            };
-            format!("{color_code}{glyph} {prose}{RESET}", glyph = self.glyph())
-        } else {
-            let label = match self {
-                Outcome::Pass => "PASS",
-                Outcome::NoProgress => "WARN",
-                Outcome::Diverged => "FAIL",
-            };
-            format!("{label}: {prose}")
-        }
-    }
-}
-
-pub const BLUE_FOREGROUND: &str = "\x1b[34m";
-pub const BLUE_BACKGROUND: &str = "\x1b[44m";
-pub const BOLD: &str = "\x1b[1m";
-pub const DIM: &str = "\x1b[2m";
-pub const GREEN: &str = "\x1b[32m";
-pub const RED: &str = "\x1b[31m";
-pub const YELLOW: &str = "\x1b[33m";
-pub const RESET: &str = "\x1b[0m";
-
-/// True when stderr points at an interactive terminal and
-/// is therefore safe to emit ANSI escape codes to.
-pub fn stderr_supports_color() -> bool {
-    std::io::stderr().is_terminal()
-}
-
-/// Owns the terminal-capability flag and all presentation
-/// methods for the `simulate` command. One instance per
-/// invocation; methods borrow `&self`.
-pub struct Reporter {
-    color: bool,
-}
-
-impl Default for Reporter {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl Reporter {
-    pub fn new() -> Self {
-        Self {
-            color: stderr_supports_color(),
-        }
-    }
-
-    pub fn banner(&self, protocol: &str, info: &[(&str, &str)]) {
-        BannerPrinter::new(protocol, info).print();
-    }
-
-    pub fn config_summary(&self, index: usize, total: usize, config: &SimulationConfig) {
-        let heading = match (total > 1, config.name.as_deref()) {
-            (true, Some(name)) => Some(format!("Simulation [{index}/{total}]: {name}")),
-            (true, None) => Some(format!("Simulation [{index}/{total}]")),
-            (false, Some(name)) => Some(format!("Simulation: {name}")),
-            (false, None) => None,
-        };
-
-        println!();
-        if let Some(heading) = heading {
-            if self.color {
-                println!("{BOLD}{heading}{RESET}");
-            } else {
-                println!("{heading}");
-            }
-        }
-        println!("{}", table::render(ConfigRow::for_config(config)));
-    }
-
-    /// Execute one simulation: spinner → run → badge → per-replica
-    /// report. Returns the outcome, the suite-level row, and the
-    /// `RunResult` so the caller can both track suite-wide aggregates
-    /// and emit structured results to a file.
-    pub async fn run(
-        &self,
-        config: SimulationConfig,
-    ) -> Result<(Outcome, SuiteRow, RunResult<SimulationConfig>)> {
-        let run_name = config.name.clone().unwrap_or_else(|| "unnamed".into());
-        let committee_size = config.committee_size;
-        let duration_secs = config.duration_secs;
-
-        let spinner = self.start_spinner();
-        let result = tokio::task::spawn_blocking(move || SimulationRunner::new(config).run())
-            .await
-            .map_err(|error| eyre!("Simulation task panicked: {error}"))?;
-        self.finish_spinner(spinner);
-
-        let outcome = result.outcome;
-        let commit_counts = result.commit_count_per_replica();
-        self.render_run(&result, duration_secs, outcome, &commit_counts);
-        println!();
-
-        let suite_row = SuiteRow::new(
-            &run_name,
-            committee_size,
-            duration_secs,
-            outcome,
-            &commit_counts,
-        );
-        Ok((outcome, suite_row, result))
-    }
-
-    fn render_run(
-        &self,
-        result: &RunResult<SimulationConfig>,
-        duration_secs: u64,
-        outcome: Outcome,
-        commit_counts: &[usize],
-    ) {
-        self.run_badge(outcome);
-
-        let rows = ReplicaRow::for_result(result, duration_secs);
-
-        // Collapse to a single-line summary in the happy path when every
-        // replica committed the same leaders and nothing is noteworthy.
-        let uniform_commits = commit_counts
-            .first()
-            .map(|first| commit_counts.iter().all(|c| c == first))
-            .unwrap_or(true);
-        let aggregate = AggregateMetrics::new(&result.metrics);
-        if outcome != Outcome::Diverged && uniform_commits && !aggregate.any_missing_blocks() {
-            let duration = Duration::from_secs(duration_secs);
-            let committed = commit_counts.first().copied().unwrap_or_default();
-            let rate = match aggregate.leader_commits_per_second(duration) {
-                Some(r) => format!("{r:.1} commits/s"),
-                None => "— commits/s".into(),
-            };
-            let mut headline = Vec::new();
-            if let (Some(p50), Some(p90)) = (aggregate.p50_latency_ms(), aggregate.p90_latency_ms())
-            {
-                headline.push(format!("p50 {p50:.0} ms · p90 {p90:.0} ms"));
-            }
-            if let Some(tps) = aggregate.committed_tps(duration) {
-                headline.push(format!("{tps:.0} TPS"));
-            }
-            self.run_summary_line(&headline.join(" · "), committed, &rate);
-            return;
-        }
-
-        self.replicas_table(rows);
-    }
-
-    fn start_spinner(&self) -> Option<ProgressBar> {
-        if !self.color {
-            return None;
-        }
-        println!();
-        let spinner = ProgressBar::new_spinner();
-        spinner.set_style(
-            ProgressStyle::with_template(" {spinner} Simulating… {elapsed}")
-                .unwrap_or_else(|_| ProgressStyle::default_spinner()),
-        );
-        spinner.enable_steady_tick(Duration::from_millis(100));
-        Some(spinner)
-    }
-
-    fn finish_spinner(&self, spinner: Option<ProgressBar>) {
-        if let Some(spinner) = spinner {
-            spinner.finish_and_clear();
-        }
-    }
-
-    fn run_badge(&self, outcome: Outcome) {
-        println!("{}", outcome.badge(self.color));
-    }
-
-    fn run_summary_line(&self, headline: &str, committed: usize, rate: &str) {
-        if headline.is_empty() {
-            println!("  {committed} commits, {rate}");
-        } else {
-            println!("  {headline} ({committed} commits, {rate})");
-        }
-    }
-
-    fn replicas_table(&self, rows: Vec<ReplicaRow>) {
-        println!("{}", table::render(rows));
-    }
-
-    pub fn suite_summary(&self, rows: &[SuiteRow]) {
-        println!();
-        if self.color {
-            println!("{BOLD}Suite summary{RESET}");
-        } else {
-            println!("Suite summary");
-        }
-        println!("{}", table::render(rows.iter().cloned()));
-    }
-}
